requirements.txt
backcall==0.1.0
decorator==4.3.2
Django==2.1.5
djangorestframework==3.9.1
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.8
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
